requirements.txt
Django==2.1.5
django-oauth-plus==2.2.9
django-oauth2-provider==0.2.6.1
django-rest-framework==0.1.0
djangorestframework==3.9.0
djangorestframework-oauth==1.1.0
httplib2==0.12.0
oauth2==1.9.0.post1
pytz==2018.9
shortuuid==0.5.0
